fetch bridge status server side
